api/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
Werkzeug==2.3.7
